refactor: move logic and functionality to StopWatch component
